+++ server/datastore/mysql/migrations/data/20170314151620_FixCentosLabel.go
@@ -1,0 +1,33 @@
+package data
+
+import (
+	"database/sql"
+
+	"github.com/kolide/kolide/server/kolide"
+)
+
+func init() {
+	MigrationClient.AddMigration(Up_20170314151620, Down_20170314151620)
+}
+
+func Up_20170314151620(tx *sql.Tx) error {
+	// Fix for osquery not correctly reporting platform for CentOS6
+	label_query := `select 1 from os_version where platform = 'centos' or name like '%centos%'`
+	sql := `
+                UPDATE labels
+                SET query = ?, platform = ''
+                WHERE name = 'CentOS Linux' AND label_type = ?
+`
+
+	_, err := tx.Exec(sql, label_query, kolide.LabelTypeBuiltIn)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func Down_20170314151620(tx *sql.Tx) error {
+	// Not reversible
+	return nil
+}